move navigasjonsknapper
